Nuttx baseline 6.18

+++ fs/fs_fdopen.c
@@ -168,7 +168,7 @@ FAR struct file_struct *fs_fdopen(int fd, int oflags, FAR _TCB *tcb)
 #endif
     }
 
-  /* The descriptor is in a valid range to file descriptor... do the read */
+  /* The descriptor is in a valid range to file descriptor... perform some more checks */
 
 #if CONFIG_NFILE_DESCRIPTORS > 0
   else
@@ -237,7 +237,7 @@ FAR struct file_struct *fs_fdopen(int fd, int oflags, FAR _TCB *tcb)
            */
 
           stream->fs_filedes = fd;
-          stream->fs_oflags  = oflags;
+          stream->fs_oflags  = (uint16_t)oflags;
 
           sem_post(&slist->sl_sem);
           return stream;
